feat(mesure): handle etat
