--- packages/app/src/components/MesureCreateOrEdit/style.js
@@ -1,5 +1,0 @@
-const ServiceMesureCreateStyle = {
-  bg: "blue",
-};
-
-export { ServiceMesureCreateStyle };
